Use GetTargetPathWithTargetPlatformMoniker and create two stage GetPackageContents
Description:
Change target used to determine if nugetizer is installed to GetTargetPathWithTargetPlatformMoniker.
This is needed for native projects (.vcxproj) that don't return an item list (so metadata can be attached)
in GetTargetPath. Instead use the target that is expected based on the metadata injection.

Also remove dependency on GetPackageTargetPath from GetPackageContents by creating a second stage,
GetFinalPackageContents. This allows consumers to potienitally adjust package metadata based on the
contents of the package. A simple example of this may be to tag the version of the package with the timestamp
of the most recently modified package member, etc.