Avoid SxS issues with NuGet tasks by IL merging
During a build, it's possible that a different version of the NuGet
assemblies are loaded by the MSBuild process (i.e. from the MSBuild
extensions path), potentially causing this package to break.

So instead of including the NuGet (and Newtonsoft.Json) assemblies
as external references, we merge them with the main tasks assemblies,
thereby isolating our tasks from those deps and ensuring that the
targets will behave as tested and shipped regardless of the installed
version of NuGet.