Keep $(Out) property default where it can be shared among targets
This property was improperly moved to build.proj, therefore causing the
projects to stop having a default value for it, and causing F5 scenarios
to break for the VS extension authoring, which needs to package the .nupkg
from the $(Out) directory for deployment to the experimental instance.

This did work fine for CI builds, since the Out property is explicitly
passed down when building solutions and projects.