Fail the build if the right nugetizer nuget package cannot be deployed/packaged
This makes far easier to detect cases where the nugetizer package doesn't exist,
and later on template unfolding will fail. It won't happen for CI builds, since
the right builds happen in the right order, but it's quite easy to make this
mistake for developer builds while iterating on the extension.