Implement a better approach than ILMerge for tasks
Instead of ILMerge, use embedded resources for the dependent assemblies,
so that on AppDomain.AssembylResolve, we'll fetch those assemblies from
the embedded resources. This guarantees that if the right binding redirects
are in place (i.e. for nuget, json.net, etc.), we'll "just work" with those
newer version, and if they are not and assembly loading fails, we'll use the
version we built against. So it's the best of both worlds.

The Fody.Costura nuget package will perform both, automatically embed the
resources, and while this makes the nuget package a bit larger, it guarantees
that it will work in MSBuild 14 as well as 15, even if the tooling for VS
only supports VS2017.