Fail on Pack if duplicate PackagePaths are found
Fixes #4554

Rendering all the source files that end up causing the duplicate was more complicated
to do in a useful manner that preserved the IDE as well as text logging readable, so
I opted to just mention the PackagePath instead. Users can easily bump up the logging
to Detailed or Diagnostic (or even better, use the MSBuild Structured Log Viewer ;))
to get a precise list of inputs to the task and see clearly which ones are the dupes.

We should still provide a hook for them to de-duplicate easily if so desired. See #5167