Added comment explaining PackOnBuildFile used by the CPS project
Under CPS, we didn't find a way to invoke a specific MSBuild target by name.

The Create Package command, under the packaging project, we just turn it into
a regular Build (see 678570eb), and use that signal file to make the Build add
Pack to its BuildDependsOn, so that it runs automatically after Build.