Improve compatibility with SDK Pack by using %(Pack) metadata
NuGetizer relies on a %(IncludeInPackage) metadata item on ProjectReference, Content and None items in order to determine if they should be processed while packing.

This represents an unnecessary migration burden on authors. We should unify on %(Pack) since it's shorter and explicit enough since it's item metadata and not a global property.

We should do this renaming in a backwards compatible way with previous NuGetizer values, so that existing nugetized projects continue to work unchanged.

Fixes https://github.com/NuGet/Home/issues/5319