Add support for Linked files as PackageFile
In the IDE it's trivial to add linked files so they end up in specified locations.

This makes it easy to add linked files by automatically using the %(Link) metadata if present
on items that don't already have a %(TargetPath) assigned.

With the addition of PackageFile as an available item name/build action, supporting
linked package files makes this very easy to use.

Fixes https://github.com/NuGet/Home/issues/5327