When items are copied to output directory, add to package from the target path
Instead of adding them to the package from the original source path. This makes
it possible for targets that tweak the outputs during build to include those
modified outputs in the package.

Concrete use case in our own NuGet.Build.Packaging.Version.props which is modified
during build to reflect the actual nugetizer version being built.