Fix intellisense breaks
We were previously ILRepack'ing to the intermediate assembly
unconditionally. This means that we'd ILRepack an already
ILRepack'ed assembly, and that failed and caused intellisense
to not work at all.

So instead, we write a flag file that tells us we have already
ilrepack'ed, and use that as the outputs of the targets.

We don't add it as a FileWrites since otherwise incremental
clean would interfere.

Also, there was a duplicate reference to the ThisAssembly.Project
nuget package that is already imported in the shared targets,
and it was also interferring with intellisense.