Implement workaround for 15.3 missing prerequisite installation
It's been confirmed that there is an issue where the prerequisites are not
installed automatically. This may be fixed by 15.3 RTW, but an official
workaround is to make the extension All Users (require Admin).

We take the chance to also install to a nice well-known location :)

Fixes https://github.com/NuGet/Home/issues/5383